Implemented feature using PersonalRatingsDated object
I've implemented the feature by creating a new PersonalRatingsDated object which is  the same as PersonalRatings but with an extra variable to store the date.
A List<PersonalRatingsDated> is converted to a JSON string and passed to the model. Thymeleaf doesn't escape the '&quot', so they must be replaced before the JSON can be parsed.

This implementation works, however, I'm working on a better implementation which would be to create a Map<String,PersonalRating> to map each PersonalRating with a date. I still need to figure out how to parse the JSON string this would create.